Separate llm vs cli tests
